+++ CHANGELOG.md
@@ -5,7 +5,8 @@ This application follows the [Semantic Versioning standard](https://semver.org/)
 ## Version 0.1.1 (2022-xx-xx)
 
 ### General
-- `From<u16>` for `ApiErrorKind` and `From<ApiErrorKind>` for `u16`.
+- Implement `From<u16>` for `ApiErrorKind` and `From<ApiErrorKind>` for `u16`.
+- Implement `Display` for `ApiErrorKind`.
 
 ## Version 0.1.0 (2021-12-22)